[#93] Fix memory leak issue + unit test
