--- components/withinBuilding/styles.js
@@ -1,9 +1,0 @@
-const styles = StyleSheet.create({
-  userFinalLoc: {
-    bottom: 80,
-    backgroundColor: 'pink',
-    position: 'absolute',
-  },
-});
-
-export default styles;
